Support KEYBOARD and CONSUMER modes of HID on Feather M4 Express
