+++ kmk/common/consts.py
@@ -5,10 +5,33 @@ class HIDReportTypes:
     SYSCONTROL = 4
 
 
+class HIDUsage:
+    KEYBOARD = 0x06
+    MOUSE = 0x02
+    CONSUMER = 0x01
+    SYSCONTROL = 0x80
+
+
+class HIDUsagePage:
+    CONSUMER = 0x0C
+    KEYBOARD = MOUSE = SYSCONTROL = 0x01
+
+
+# Currently only used by the CircuitPython HIDHelper because CircuitPython
+# actually enforces these limits with a ValueError. Unused on PyBoard because
+# we can happily send full reports there and it magically works.
+HID_REPORT_SIZES = {
+    HIDReportTypes.KEYBOARD: 8,
+    HIDReportTypes.MOUSE: 4,
+    HIDReportTypes.CONSUMER: 2,
+    HIDReportTypes.SYSCONTROL: 8,  # TODO find the correct value for this
+}
+
+
 HID_REPORT_STRUCTURE = bytes([
     # Regular keyboard
-    0x05, 0x01,  # Usage Page (Generic Desktop)
-    0x09, 0x06,  # Usage (Keyboard)
+    0x05, HIDUsagePage.KEYBOARD,  # Usage Page (Generic Desktop)
+    0x09, HIDUsage.KEYBOARD,  # Usage (Keyboard)
     0xA1, 0x01,  # Collection (Application)
     0x85, HIDReportTypes.KEYBOARD,  #   Report ID (1)
     0x05, 0x07,  #   Usage Page (Keyboard)
@@ -39,8 +62,8 @@ HID_REPORT_STRUCTURE = bytes([
     0x91, 0x01,  #   Output (Constant)
     0xC0,        # End Collection
     # Regular mouse
-    0x05, 0x01,  # Usage Page (Generic Desktop)
-    0x09, 0x02,  # Usage (Mouse)
+    0x05, HIDUsagePage.MOUSE,  # Usage Page (Generic Desktop)
+    0x09, HIDUsage.MOUSE,  # Usage (Mouse)
     0xA1, 0x01,  # Collection (Application)
     0x09, 0x01,  #   Usage (Pointer)
     0xA1, 0x00,  #   Collection (Physical)
@@ -73,8 +96,8 @@ HID_REPORT_STRUCTURE = bytes([
     0xC0,        #   End Collection
     0xC0,        # End Collection
     # Consumer ("multimedia") keys
-    0x05, 0x0C,        # Usage Page (Consumer)
-    0x09, 0x01,        # Usage (Consumer Control)
+    0x05, HIDUsagePage.CONSUMER,  # Usage Page (Consumer)
+    0x09, HIDUsage.CONSUMER,  # Usage (Consumer Control)
     0xA1, 0x01,        # Collection (Application)
     0x85, HIDReportTypes.CONSUMER,  # Report ID (n)
     0x75, 0x10,        #   Report Size (16)
@@ -86,8 +109,8 @@ HID_REPORT_STRUCTURE = bytes([
     0x81, 0x00,        #   Input (Data,Array,Abs,No Wrap,Linear,Preferred State,No Null Position)
     0xC0,              # End Collection
     # Power controls
-    0x05, 0x01,        # Usage Page (Generic Desktop Ctrls)
-    0x09, 0x80,        # Usage (Sys Control)
+    0x05, HIDUsagePage.SYSCONTROL,  # Usage Page (Generic Desktop Ctrls)
+    0x09, HIDUsage.SYSCONTROL,  # Usage (Sys Control)
     0xA1, 0x01,        # Collection (Application)
     0x85, HIDReportTypes.SYSCONTROL,  # Report ID (n)
     0x75, 0x02,        #   Report Size (2)
